created a separate api folder for user auth routes and zillow api calls and fixed all axios calls in the frontend
